+++ sdk/storage/azure-storage-blobs/test/blob_service_client_test.cpp
@@ -11,7 +11,7 @@ namespace Azure { namespace Storage { namespace Blobs { namespace Models {
 
   bool operator==(const BlobRetentionPolicy& lhs, const BlobRetentionPolicy& rhs)
   {
-    if (lhs.Enabled != rhs.Enabled)
+    if (lhs.IsEnabled != rhs.IsEnabled)
     {
       return false;
     }
@@ -35,7 +35,7 @@ namespace Azure { namespace Storage { namespace Blobs { namespace Models {
 
   bool operator==(const BlobStaticWebsite& lhs, const BlobStaticWebsite& rhs)
   {
-    if (lhs.Enabled != rhs.Enabled)
+    if (lhs.IsEnabled != rhs.IsEnabled)
     {
       return false;
     }
@@ -170,7 +170,7 @@ namespace Azure { namespace Storage { namespace Test {
     auto properties = *ret;
     auto logging = properties.Logging;
     EXPECT_FALSE(logging.Version.empty());
-    if (logging.RetentionPolicy.Enabled)
+    if (logging.RetentionPolicy.IsEnabled)
     {
       EXPECT_TRUE(logging.RetentionPolicy.Days.HasValue());
     }
@@ -178,7 +178,7 @@ namespace Azure { namespace Storage { namespace Test {
     if (hourMetrics.IsEnabled)
     {
       EXPECT_FALSE(hourMetrics.Version.empty());
-      if (hourMetrics.RetentionPolicy.Enabled)
+      if (hourMetrics.RetentionPolicy.IsEnabled)
       {
         EXPECT_TRUE(hourMetrics.RetentionPolicy.Days.HasValue());
       }
@@ -187,13 +187,13 @@ namespace Azure { namespace Storage { namespace Test {
     if (minuteMetrics.IsEnabled)
     {
       EXPECT_FALSE(minuteMetrics.Version.empty());
-      if (minuteMetrics.RetentionPolicy.Enabled)
+      if (minuteMetrics.RetentionPolicy.IsEnabled)
       {
         EXPECT_TRUE(minuteMetrics.RetentionPolicy.Days.HasValue());
       }
     }
     auto deleteRetentionPolicy = properties.DeleteRetentionPolicy;
-    if (deleteRetentionPolicy.Enabled)
+    if (deleteRetentionPolicy.IsEnabled)
     {
       EXPECT_TRUE(deleteRetentionPolicy.Days.HasValue());
     }
@@ -216,22 +216,22 @@ namespace Azure { namespace Storage { namespace Test {
     properties.Logging.Delete = !properties.Logging.Delete;
     properties.Logging.Read = !properties.Logging.Read;
     properties.Logging.Write = !properties.Logging.Write;
-    properties.Logging.RetentionPolicy.Enabled = true;
+    properties.Logging.RetentionPolicy.IsEnabled = true;
     properties.Logging.RetentionPolicy.Days = 3;
 
     properties.HourMetrics.IsEnabled = true;
-    properties.HourMetrics.RetentionPolicy.Enabled = true;
+    properties.HourMetrics.RetentionPolicy.IsEnabled = true;
     properties.HourMetrics.RetentionPolicy.Days = 4;
     properties.HourMetrics.IncludeApis = true;
 
     properties.MinuteMetrics.IsEnabled = true;
-    properties.MinuteMetrics.RetentionPolicy.Enabled = true;
+    properties.MinuteMetrics.RetentionPolicy.IsEnabled = true;
     properties.MinuteMetrics.RetentionPolicy.Days = 4;
     properties.MinuteMetrics.IncludeApis = true;
 
     properties.DefaultServiceVersion = Blobs::Details::ApiVersion;
 
-    properties.StaticWebsite.Enabled = true;
+    properties.StaticWebsite.IsEnabled = true;
     properties.StaticWebsite.IndexDocument = "index.html";
     properties.StaticWebsite.ErrorDocument404Path = "404.html";
     properties.StaticWebsite.DefaultIndexDocumentPath.Reset();
@@ -251,7 +251,7 @@ namespace Azure { namespace Storage { namespace Test {
     corsRule.MaxAgeInSeconds = 20;
     properties.Cors.emplace_back(corsRule);
 
-    properties.DeleteRetentionPolicy.Enabled = true;
+    properties.DeleteRetentionPolicy.IsEnabled = true;
     properties.DeleteRetentionPolicy.Days = 7;
 
     EXPECT_NO_THROW(m_blobServiceClient.SetProperties(properties));